03.02.2026 21.21

+++ template.cpp
@@ -32,7 +32,7 @@ int32_t main() {
 	ios_base::sync_with_stdio(false);
 	cin.tie(nullptr); file();
 
-	ll t = 1; //cin >> t;
+	int t = 1; //cin >> t;
 	while (t--) testcase();
 
 	return 0;
